// src/app/model/spp-register.ts
import {Group} from './group.class';
import {Role} from './role.class';

export class SppRegister {
  accucode: string;
  againstcode: string;
  appecode: string;
  bcvnd: true;
  begin_setnum: string;
  bregicodetha: string;
  casecode: string;
  casename: string;
  casetype: string;
  checkhasregi: true;
  editable: string;
  finishdate: Date;
  fromdate: Date;
  fullname: string;
  indate: Date;
  lawcode: string;
  lawyer: true;
  regiclosed: string;
  regicode: string;
  remark: string;
  setnum: string;
  spcindate: Date;
  spcsetnum: string;
  special: true;
  sppid: string;
  sppname: string;
  sppspc: string;
  status: string;
  statusname: string;
  statusreason: string;
  strmdeciid: string;
  tgvpl: true;
  todate: Date;
  transfer: string;
  transferout: string;
  unspecial: string;
  updnum: string;
  userfor: string;
  userforname: string;
  evidence: string;
  num_of_pen: number;
  registerdeciid: number;
}